Create flow for monte carlo method, strip main back to allow  for integration of diferent learning flows
