+++ src/control/setup/__init__.py
@@ -1,0 +1,3 @@
+from .pre_run_tasks import pre_run_calculations_tasks
+
+__all__ = ["pre_run_calculations_tasks"]
